+++ src/array_subset/iterators/chunks_iterator.rs
@@ -6,8 +6,9 @@ use rayon::iter::{
 };
 
 use crate::{
-    array::{chunk_shape_to_array_shape, ArrayIndices},
+    array::ArrayIndices,
     array_subset::{ArraySubset, IncompatibleDimensionalityError},
+    metadata::chunk_shape_to_array_shape,
 };
 
 use super::{